+++ hydra-node/src/Hydra/API/ServerOutput.hs
@@ -69,7 +69,6 @@ data ClientMessage tx
   | PostTxOnChainFailed {postChainTx :: PostChainTx tx, postTxError :: PostTxError tx}
   | RejectedInputBecauseUnsynced {clientInput :: ClientInput tx, drift :: NominalDiffTime}
   | SideLoadSnapshotRejected {clientInput :: ClientInput tx, requirementFailure :: SideLoadRequirementFailure tx}
-  | SyncedStatusReport {chainSlot :: ChainSlot, chainTime :: UTCTime, drift :: NominalDiffTime, synced :: SyncedStatus}
   deriving stock (Eq, Show, Generic)
 
 instance IsChainState tx => ToJSON (ClientMessage tx) where